Only instantiate db handlers once
