+++ trin-core/src/utils/db.rs
@@ -1,11 +1,8 @@
-use std::sync::Arc;
 use std::{env, fs};
 
 use directories::ProjectDirs;
 use discv5::enr::NodeId;
 use rocksdb::{Options, DB};
-
-use crate::portalnet::storage::{DistanceFunction, PortalStorage, PortalStorageConfig};
 
 const TRIN_DATA_ENV_VAR: &str = "TRIN_DATA_PATH";
 
@@ -38,18 +35,4 @@ pub fn setup_overlay_db(node_id: NodeId) -> DB {
     let mut db_opts = Options::default();
     db_opts.create_if_missing(true);
     DB::open(&db_opts, data_path).unwrap()
-}
-
-pub fn setup_portal_storage(node_id: NodeId, kb: u32) -> PortalStorage {
-    let rocks_db = PortalStorage::setup_rocksdb(node_id).unwrap();
-    let meta_db = PortalStorage::setup_sqlite(node_id).unwrap();
-
-    let storage_config = PortalStorageConfig {
-        storage_capacity_kb: (kb / 4) as u64,
-        node_id,
-        distance_function: DistanceFunction::Xor,
-        db: Arc::new(rocks_db),
-        meta_db: Arc::new(meta_db),
-    };
-    PortalStorage::new(storage_config).unwrap()
 }
