+++ trin-state/src/network.rs
@@ -8,9 +8,10 @@ use tokio::sync::RwLock;
 use trin_core::portalnet::{
     discovery::Discovery,
     overlay::{OverlayConfig, OverlayProtocol, OverlayRequestError},
+    storage::{PortalStorage, PortalStorageConfig},
     types::messages::{PortalnetConfig, ProtocolId},
 };
-use trin_core::utils::db::{setup_overlay_db, setup_portal_storage};
+use trin_core::utils::db::setup_overlay_db;
 use trin_core::utp::stream::UtpListener;
 
 use crate::{content_key::StateContentKey, trie::TrieDB};
@@ -26,16 +27,15 @@ impl StateNetwork {
     pub async fn new(
         discovery: Arc<Discovery>,
         utp_listener: Arc<RwLock<UtpListener>>,
-        storage_kb: u32,
+        storage_config: PortalStorageConfig,
         portal_config: PortalnetConfig,
     ) -> Self {
-        // no bueno
+        // todo: revisit triedb location
         let db = setup_overlay_db(NodeId::random());
         let triedb = TrieDB::new(Arc::new(db));
         let trie = EthTrie::new(Arc::new(triedb));
 
-        let node_id = discovery.local_enr().node_id();
-        let storage = Arc::new(setup_portal_storage(node_id, storage_kb));
+        let storage = Arc::new(PortalStorage::new(storage_config).unwrap());
         let config = OverlayConfig {
             bootnode_enrs: portal_config.bootnode_enrs.clone(),
             ..Default::default()
